app/src/main: add new RestClient class
This class support calling the Gemini model by invoking the
POST method. User need to create an object of this class
passing a timeout value, this is the timeout value of
the HTTP request. The class provide two callback that user
can implements which will get called once an event occur
(onSuccess, onFailure). The POST method accepts message
as input token and user can specify the maximum generated
output text from the model.